requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos[dbt-snowflake]==1.0.3
protobuf==3.20.0
soda-core-snowflake==3.0.45
apache-airflow-providers-snowflake

